+++ numina/core/recipeinout.py
@@ -22,9 +22,10 @@ Recipe Input
 """
 
 from six import with_metaclass
+import yaml
 
 from .metaclass import RecipeInputType, RecipeResultType
-
+from numina.store.dump import dump
 
 class RecipeInOut(object):
 
@@ -105,6 +106,12 @@ class ErrorRecipeResult(BaseRecipeResult):
         fmt = "%s(errortype=%r, message='%s')"
         return fmt % (sclass, self.errortype, self.message)
 
+    def store(self, where):
+        with open(where.result, 'w+') as fd:
+            yaml.dump(where.result, fd)
+
+        return where.result
+
 
 class RecipeResult(with_metaclass(RecipeResultType, RecipeInOut, BaseRecipeResult)):
 
@@ -114,6 +121,19 @@ class RecipeResult(with_metaclass(RecipeResultType, RecipeInOut, BaseRecipeResul
         for key, val in self.stored().items():
             full.append('%s=%r' % (key, val))
         return '%s(%s)' % (sclass, ', '.join(full))
+
+    def store(self, where):
+
+        saveres = {}
+        for key, pc in self.stored().items():
+            val = getattr(self, key)
+            where.destination = pc.dest
+            saveres[key] = dump(pc.type, val, where)
+
+        with open(where.result, 'w+') as fd:
+            yaml.dump(saveres, fd)
+
+        return where.result
 
 
 class define_result(object):
