+++ numina/store/__init__.py
@@ -1,4 +1,3 @@
 
-from .dump import dump
 from .load import load
