+++ numina/store/default.py
@@ -17,66 +17,12 @@
 # along with Numina.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-"""Register basic types with dump."""
+"""Two basic functions to store frames and arrays"""
 
-from __future__ import print_function
 
 import warnings
 
 import numpy
-import yaml
-
-from numina.core import RecipeResult
-from numina.core import ErrorRecipeResult
-from numina.core import DataFrameType, DataProductType
-from numina.core.types import PlainPythonType
-from numina.core.products import ArrayType
-from numina.core import DataFrame
-
-from .dump import dump
-
-
-@dump.register(ErrorRecipeResult)
-def _(tag, obj, where):
-    with open(where.result, 'w+') as fd:
-        yaml.dump(where.result, fd)
-
-    return where.result
-
-
-@dump.register(RecipeResult)
-def _(tag, obj, where):
-
-    saveres = {}
-    for key, pc in obj.stored().items():
-        val = getattr(obj, key)
-        where.destination = pc.dest
-        saveres[key] = dump(pc.type, val, where)
-
-    with open(where.result, 'w+') as fd:
-        yaml.dump(saveres, fd)
-
-    return where.result
-
-
-@dump.register(DataProductType)
-def _(tag, obj, where):
-    return obj
-
-
-@dump.register(PlainPythonType)
-def _(tag, obj, where):    
-    return obj
-
-
-@dump.register(ArrayType)
-def _(tag, obj, where):
-    return dump_numpy_array(obj, where)
-
-
-@dump.register(DataFrameType)
-def _(tag, obj, where):
-    return dump_dataframe(obj, where)
 
 
 def dump_dataframe(obj, where):
@@ -107,19 +53,6 @@ def dump_numpy_array(obj, where):
     return filename
 
 
-dump.register(numpy.ndarray, dump_numpy_array)
-
-dump.register(DataFrame, dump_dataframe)
-
-
-@dump.register(list)
-def _(tag, obj, where):
-    return [dump(tag, o, where) for o in obj]
-
-
-# FIXME: this is very convoluted
-from .defaultl import load_cli_storage as other
-
-
+# FIXME: remove this function together with init_dump_backends
 def load_cli_storage():
-    return other()
+    pass
