+++ numina/store/dump.py
@@ -1,4 +1,22 @@
-from __future__ import print_function
+#
+# Copyright 2010-2015 Universidad Complutense de Madrid
+#
+# This file is part of Numina
+#
+# Numina is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Numina is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Numina.  If not, see <http://www.gnu.org/licenses/>.
+#
+
 
 
 try:
@@ -6,8 +24,29 @@ try:
 except ImportError:
     from singledispatch import singledispatch
 
+import numpy
+
+from numina.core.dataframe import DataFrame
+from .default import dump_dataframe, dump_numpy_array
 
 @singledispatch
 def dump(tag, obj, where):
-    return obj
 
+    if hasattr(tag, '__numina_dump__'):
+        return tag.__numina_dump__(obj, where)
+    else:
+        return obj
+
+# It's not clear if I need to register these three
+# functions
+
+
+@dump.register(list)
+def _(tag, obj, where):
+    return [dump(tag, o, where) for o in obj]
+
+
+dump.register(numpy.ndarray, dump_numpy_array)
+
+
+dump.register(DataFrame, dump_dataframe)
